+++ src/tests/query.rs
@@ -177,14 +177,14 @@ mod tests {
         );
 
         assert!(sql.contains("DummySchema._id"));
-        assert!(sql.contains(" FROM DummySchema"));
-
-        println!("SQL: {sql}");
-        println!("Query Joins: {:?}", query.joins);
+        #[cfg(feature = "mysql")]
+        assert!(sql.contains(" FROM `DummySchema`"));
+        #[cfg(feature = "postgres")]
+        assert!(sql.contains(" FROM \"DummySchema\""));
+        #[cfg(feature = "sqlite")]
+        assert!(sql.contains(" FROM \"DummySchema\""));
 
         let sql = Query::<DummySchema, SelectDummySchema>::joins_sql(sql, &query.joins);
-
-        println!("SQL: {sql}");
 
         assert!(sql.contains("LEFT JOIN"));
     }